# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "operationId": "UsersController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/users/{id}/status": {
      "patch": {
        "operationId": "UsersController_updateStatus_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "UAEAF Backend API",
    "description": "UAE Athletics Federation platform API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LocalizedTextDto": {
        "type": "object",
        "properties": {
          "en": {
            "type": "string",
            "description": "English text."
          },
          "ar": {
            "type": "string",
            "description": "Arabic text."
          }
        },
        "required": [
          "en",
          "ar"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "description": "Full name, recorded in both Arabic and English.",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "email": {
            "type": "string",
            "description": "Login email, must be unique."
          },
          "password": {
            "type": "string",
            "description": "Plaintext password — hashed before storage, never persisted as-is."
          },
          "roleIds": {
            "description": "Roles the account starts with. Validated against the live roles collection; an unknown or archived id rejects the whole request. Omit for an account with no access.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "personId": {
            "type": "string",
            "description": "Optional link to the federationPersonnel record this account belongs to. Verified to exist before the account is written.",
            "nullable": true
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "email": {
            "type": "string"
          },
          "roleIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "personId": {
            "type": "object",
            "nullable": true
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "Active",
              "Suspended",
              "Deactivated"
            ]
          },
          "lastLogin": {
            "type": "object",
            "nullable": true
          },
          "photoId": {
            "type": "object",
            "nullable": true
          },
          "preferredLanguage": {
            "type": "string",
            "enum": [
              "ar",
              "en"
            ],
            "nullable": true
          },
          "preferredTheme": {
            "type": "string",
            "enum": [
              "light",
              "dark"
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "roleIds",
          "accountStatus"
        ]
      },
...
      "UpdateAccountStatusDto": {
        "type": "object",
        "properties": {
          "accountStatus": {
            "type": "string",
            "description": "The account state to move to. Anything other than Active also ends every live session for that account.",
            "enum": [
              "Active",
              "Suspended",
              "Deactivated"
            ]
          }
        },
        "required": [
          "accountStatus"
        ]
      },
...
    }
  }
}